fix(divider): fixed up divider and all components in IUI that use it - Stepper refactored to be more IUI and use rem
