+++ .storybook/decorators/icons.js
@@ -8,6 +8,7 @@ import {
   faBell,
   faCaretDown,
   faCaretUp,
+  faCheck,
   faCircle,
   faCloud,
   faCopy as fasCopy,
@@ -44,6 +45,7 @@ library.add(
   faBell,
   faCaretDown,
   faCaretUp,
+  faCheck,
   faCircle,
   farCopy,
   fasCopy,
