+++ components/atoms/avatar/variables.js
@@ -1,6 +1,8 @@
 /**
  * Avatar - Variables
  */
+
+// UI
 import { COLOUR } from '../../theme/variables/colour'
 
 export const AVATAR = {
