+++ components/form/datePicker/calendar/__tests__/stories.js
@@ -64,7 +64,7 @@ const BaseComponent = (props = {}) => {
     <FormForm handleSubmit={handleSubmit(onSubmit)}>
       <DatePickerCalendar {...defaultProps} />
 
-      <Divider />
+      <Divider size='sm' />
 
       <Button content='Submit' type='submit' />
