+++ components/molecules/stepper/components/stepper.js
@@ -6,7 +6,7 @@
 import { array, func, oneOfType, string } from 'prop-types'
 
 // UI
-import { StepperItem, StepperSummary } from './'
+import { StepperItem, StepperSummary } from '../../../'
 
 // Style
 import styled from 'styled-components'
@@ -15,8 +15,9 @@ export const Stepper = ({ className, items, summary }) => {
   return (
     <StyledStepper>
       <ul>
-        {items.map(item => item.active !== false && <StepperItem key={item.id} item={item} />)}
+        {items.map(item => item.active !== false && <StepperItem item={item} key={item.id} />)}
       </ul>
+
       {summary && <StepperSummary summary={summary} />}
     </StyledStepper>
   )
@@ -26,8 +27,8 @@ const StyledStepper = styled.div`
   display: inline-block;
   > ul {
     list-style: none;
-    margin: 0;
-    padding: 0 0 0 20px;
+    margin: 0 0.5rem 1rem;
+    padding: 0;
   }
 `
